+++ node/src/main/kotlin/net/corda/notary/experimental/raft/RaftConfig.kt
@@ -1,4 +1,4 @@
-package net.corda.notary.raft
+package net.corda.notary.experimental.raft
 
 import net.corda.core.utilities.NetworkHostAndPort
