+++ node/src/main/resources/net/corda/notary/experimental/bftsmart/system.config.printf
