+++ node/src/test/resources/net/corda/notary/experimental/bftsmart/system.config.printf
